Fix many of the intcap_t bitwise operation warnings by either:
1) using the new q{Get,Set,Clear}LowPointerBits functions in QtBase
2) using __builtin_align_{up, down}
3) replacing the intptr_t with an integer type if that is the intention

+++ Source/JavaScriptCore/llint/LLIntThunks.cpp
@@ -159,7 +159,7 @@ EncodedJSValue vmEntryToNative(void* executableAddress, VM* vm, ProtoCallFrame* 
 extern "C" VMEntryRecord* vmEntryRecord(EntryFrame* entryFrame)
 {
     // The C Loop doesn't have any callee save registers, so the VMEntryRecord is allocated at the base of the frame.
-    intptr_t stackAlignment = stackAlignmentBytes();
+    unsigned stackAlignment = stackAlignmentBytes();
     intptr_t VMEntryTotalFrameSize = (sizeof(VMEntryRecord) + (stackAlignment - 1)) & ~(stackAlignment - 1);
     return reinterpret_cast<VMEntryRecord*>(reinterpret_cast<char*>(entryFrame) - VMEntryTotalFrameSize);
 }
